+++ src/ProbabilityMeasures.jl
@@ -35,6 +35,8 @@ include("univariate/discrete/binomial.jl")
 
 include("multivariate/continuous/mvnormal.jl")
 
+include("multivariate/discrete/multinomial.jl")
+
 # Export the operations commonly needed by probabilistic programs.
 
 # Core types
@@ -42,10 +44,11 @@ export AbstractProbabilityMeasure
 export VariateForm, Univariate, Multivariate
 export ValueSupport, Continuous, Discrete
 export ContinuousUnivariateMeasure, DiscreteUnivariateMeasure
-export ContinuousMultivariateMeasure
+export ContinuousMultivariateMeasure, DiscreteMultivariateMeasure
 
 # Supports
-export Support, RealLine, NonNegativeReals, RealInterval, IntegerRange, RealVectors
+export Support, RealLine, NonNegativeReals, RealInterval, IntegerRange
+export RealVectors, CountVectors
 export support, insupport
 
 # Interface
@@ -66,5 +69,6 @@ export Categorical
 export Bernoulli
 export Binomial
 export MvNormal
+export Multinomial
 
 end